multi tenancy streamlined

+++ src/abstractions/Backend.Fx/Environment/MultiTenancy/TenantService.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Linq;
+using Backend.Fx.Exceptions;
 using Backend.Fx.Logging;
 using Backend.Fx.Patterns.EventAggregation.Integration;
-using JetBrains.Annotations;
 
 namespace Backend.Fx.Environment.MultiTenancy
 {
@@ -12,23 +12,24 @@ namespace Backend.Fx.Environment.MultiTenancy
     /// </summary>
     public interface ITenantService
     {
-        ITenantIdProvider IdProvider { get; }
-        
-        TenantId CreateTenant(TenantCreationParameters param);
+        /// <summary>
+        /// The tenant service can also provide an <see cref="ITenantIdProvider"/>. Keep in mind that this instance uses a direct
+        /// database connection. When multiple microservices do not share the same database, this instance cannot be used, but must
+        /// be implemented by a client to the master tenant service, probably using a remoting technology like RESTful Service, HTTP,
+        /// gRPC or SOAP web service 
+        /// </summary>
+        ITenantIdProvider TenantIdProvider { get; }
+
+        TenantId CreateTenant(string name, string description, bool isDemonstrationTenant, string configuration = null);
         void ActivateTenant(TenantId tenantId);
         void DeactivateTenant(TenantId tenantId);
         void DeleteTenant(TenantId tenantId);
-        
+
+        Tenant[] GetTenants();
         Tenant[] GetActiveTenants();
         Tenant[] GetActiveDemonstrationTenants();
         Tenant[] GetActiveProductionTenants();
-    }
-
-    public interface ITenantIdProvider
-    {
-        TenantId[] GetActiveTenantIds();
-        TenantId[] GetActiveDemonstrationTenantIds();
-        TenantId[] GetActiveProductionTenantIds();
+        Tenant GetTenant(TenantId tenantId);
     }
 
     public class TenantService : ITenantService
@@ -37,19 +38,35 @@ namespace Backend.Fx.Environment.MultiTenancy
         private readonly IMessageBus _messageBus;
         private readonly ITenantRepository _tenantRepository;
 
-        public ITenantIdProvider IdProvider { get; }
+        public ITenantIdProvider TenantIdProvider { get; }
 
         public TenantService(IMessageBus messageBus, ITenantRepository tenantRepository)
         {
             _messageBus = messageBus;
             _tenantRepository = tenantRepository;
-            IdProvider = new TenantServiceTenantIdProvider(this);
+            TenantIdProvider = new TenantServiceTenantIdProvider(this);
         }
 
-        public TenantId CreateTenant(TenantCreationParameters param)
+        public TenantId CreateTenant(string name, string description, bool isDemonstrationTenant, string configuration = null)
         {
-            Logger.Info($"Creating tenant: {param.Name}");
-            return CreateTenant(param.Name, param.Description, param.IsDemonstrationTenant, param.Configuration);
+            Logger.Info($"Creating tenant: {name}");
+            Logger.Info($"Creating Tenant {name}");
+
+            if (string.IsNullOrWhiteSpace(name))
+            {
+                throw new ArgumentException("Value cannot be null or whitespace.", nameof(name));
+            }
+
+            if (_tenantRepository.GetTenants().Any(t => t.Name != null && t.Name.ToLowerInvariant() == name.ToLowerInvariant()))
+            {
+                throw new ArgumentException($"There is already a tenant named {name}");
+            }
+
+            var tenant = new Tenant(name, description, isDemonstrationTenant) {Configuration = configuration};
+            _tenantRepository.SaveTenant(tenant);
+            var tenantId = new TenantId(tenant.Id);
+            _messageBus.Publish(new TenantActivated(tenant.Id, tenant.Name, tenant.Description, tenant.IsDemoTenant));
+            return tenantId;
         }
 
         public void ActivateTenant(TenantId tenantId)
@@ -73,60 +90,58 @@ namespace Backend.Fx.Environment.MultiTenancy
         public void DeleteTenant(TenantId tenantId)
         {
             Logger.Info($"Deleting tenant: {tenantId}");
+            Tenant tenant = _tenantRepository.GetTenant(tenantId);
+            if (tenant.State != TenantState.Inactive)
+            {
+                throw new UnprocessableException($"Attempt to delete active tenant[{tenantId.Value}]")
+                    .AddError("You cannot delete an active tenant. Please make sure to deactivate it first.");
+            }
+
             _tenantRepository.DeleteTenant(tenantId);
+        }
+
+        public Tenant GetTenant(TenantId tenantId)
+        {
+            return _tenantRepository.GetTenant(tenantId);
+        }
+
+        public Tenant[] GetTenants()
+        {
+            var tenants = _tenantRepository.GetTenants();
+            Logger.Trace($"TenantIds: {string.Join(",", tenants.Select(t => t.ToString()))}");
+            return tenants;
         }
 
         public Tenant[] GetActiveTenants()
         {
-            var activeTenantIds = _tenantRepository
+            var activeTenants = _tenantRepository
                                   .GetTenants()
                                   .Where(t => t.State == TenantState.Active)
                                   .ToArray();
-            Logger.Trace($"Active TenantIds: {string.Join(",",activeTenantIds.Select(t => t.ToString()))}");
-            return activeTenantIds;
+            Logger.Trace($"Active TenantIds: {string.Join(",", activeTenants.Select(t => t.ToString()))}");
+            return activeTenants;
         }
 
         public Tenant[] GetActiveDemonstrationTenants()
         {
-            var activeDemonstrationTenantIds = _tenantRepository
+            var activeDemonstrationTenants = _tenantRepository
                                                .GetTenants()
                                                .Where(t => t.State == TenantState.Active && t.IsDemoTenant)
                                                .ToArray();
-            Logger.Trace($"Active Demonstration TenantIds: {string.Join(",",activeDemonstrationTenantIds.Select(t => t.ToString()))}");
-            return activeDemonstrationTenantIds;
+            Logger.Trace($"Active Demonstration TenantIds: {string.Join(",", activeDemonstrationTenants.Select(t => t.ToString()))}");
+            return activeDemonstrationTenants;
         }
 
         public Tenant[] GetActiveProductionTenants()
         {
-            var activeProductionTenantIds = _tenantRepository
+            var activeProductionTenants = _tenantRepository
                                             .GetTenants()
                                             .Where(t => t.State == TenantState.Active && !t.IsDemoTenant)
                                             .ToArray();
-            Logger.Trace($"Active Production TenantIds: {string.Join(",",activeProductionTenantIds.Select(t => t.ToString()))}");
-            return activeProductionTenantIds;
+            Logger.Trace($"Active Production TenantIds: {string.Join(",", activeProductionTenants.Select(t => t.ToString()))}");
+            return activeProductionTenants;
         }
 
-        protected virtual TenantId CreateTenant([NotNull] string name, string description, bool isDemo, string configuration)
-        {
-            Logger.Info($"Creating Tenant {name}");
-            
-            if (string.IsNullOrWhiteSpace(name))
-            {
-                throw new ArgumentException("Value cannot be null or whitespace.", nameof(name));
-            }
-
-            if (_tenantRepository.GetTenants().Any(t => t.Name != null && t.Name.ToLowerInvariant() == name.ToLowerInvariant()))
-            {
-                throw new ArgumentException($"There is already a tenant named {name}");
-            }
-
-            var tenant = new Tenant(name, description, isDemo) {Configuration = configuration};
-            _tenantRepository.SaveTenant(tenant);
-            var tenantId = new TenantId(tenant.Id);
-            _messageBus.Publish(new TenantActivated(tenant.Id, tenant.Name, tenant.Description, tenant.IsDemoTenant));
-            return tenantId;
-        }
-        
         private class TenantServiceTenantIdProvider : ITenantIdProvider
         {
             private readonly ITenantService _tenantService;
@@ -139,17 +154,17 @@ namespace Backend.Fx.Environment.MultiTenancy
             public TenantId[] GetActiveDemonstrationTenantIds()
             {
                 return _tenantService.GetActiveDemonstrationTenants()
-                                       .Select(t => new TenantId(t.Id))
-                                       .ToArray();
+                                     .Select(t => new TenantId(t.Id))
+                                     .ToArray();
             }
 
             public TenantId[] GetActiveProductionTenantIds()
             {
                 return _tenantService.GetActiveProductionTenants()
-                                       .Select(t => new TenantId(t.Id))
-                                       .ToArray();
+                                     .Select(t => new TenantId(t.Id))
+                                     .ToArray();
             }
-            
+
             public TenantId[] GetActiveTenantIds()
             {
                 return _tenantService.GetActiveTenants()
